feat: article-meta changed component name to date-publication and updated code according to PR
