+++ packages/date-publication/.npmignore
